agentic-kahoot-demo/requirements.txt
pymupdf
crewai>=0.140.0
crewai-tools[mcp]>=0.42.0
python-dotenv>=1.0.0
pydantic>=2.0.0
mcp>=1.0.0
chromadb>=0.4.24
docling
openai
numpy<2
